authorization implemented

+++ src/api/notes/handler.js
@@ -19,7 +19,9 @@ class NotesHandler {
 
       const {title = 'untilted', body, tags} = request.payload;
 
-      const noteId = await this._service.addNote({title, body, tags});
+      const {id: credentialId } = request.auth.credentials;
+
+      const noteId = await this._service.addNote({title, body, tags, owner: credentialId});
 
       const response = h.response({
         status: 'success',
@@ -58,9 +60,10 @@ class NotesHandler {
     }
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
-    console.error(notes);
+  async getNotesHandler(request, h) {
+    const {id: credentialId } = request.auth.credentials;
+
+    const notes = await this._service.getNotes(credentialId);
     return {
       status: 'success',
       data: {
@@ -73,6 +76,10 @@ class NotesHandler {
     
     try {
       const { id } = request.params;
+
+      const {id: credentialId } = request.auth.credentials;
+      
+      await this._service.verifyNoteOwner(id, credentialId);
 
       const note = await this._service.getNoteById(id);
 
@@ -113,6 +120,10 @@ class NotesHandler {
       
       const { id } = request.params;
 
+      const {id: credentialId } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentialId);
+
       await this._service.editNoteById(id, request.payload);
 
       return {
@@ -146,12 +157,39 @@ class NotesHandler {
   }
 
   async deleteNoteByIdHandler(request, h) {
-    const { id } = request.params;
-    await this._service.deleteNoteById(id);
-    return {
-      status: 'success',
-      message: 'Catatan berhasil dihapus',
+    try {
+      const { id } = request.params;
+      const {id: credentialId } = request.auth.credentials;
+  
+      await this._service.verifyNoteOwner(id, credentialId);
+  
+      await this._service.deleteNoteById(id);
+      
+      return {
+        status: 'success',
+        message: 'Catatan berhasil dihapus',
+      }
+    } catch (error) {
+      if (error instanceof ClientError) {
+        const response = h.response({
+          status: 'fail',
+          message: error.message,
+        });
+        response.code(error.statusCode);
+        return response;
+      }
+
+       // Server ERROR!
+      const response = h.response({
+        status: 'error',
+        message: 'Maaf, terjadi kegagalan pada server kami.',
+      });
+      response.code(500);
+      console.error(error);
+      return response;
     }
+
+
   }
 }
